refactor(runtime): narrow compaction input ownership
