+++ packages/runtime/src/__tests__/history-compact-artifacts.test.ts
@@ -3,7 +3,7 @@ import { Buffer } from 'node:buffer';
 import { describe, test } from 'node:test';
 import type { RuntimeEvent } from '@maka/core/runtime-event';
 import { buildHistoryCompactBlockFromSummary } from '../context-budget.js';
-import type { HistoryCompactWriteInput } from '../ai-sdk-backend.js';
+import type { HistoryCompactWriteInput } from '../ai-sdk-compaction-contract.js';
 import {
   loadHistoryCompactBlocksFromArtifacts,
   persistHistoryCompactBlocksToArtifacts,
